Refractoring code for app import compression
