+++ corehq/apps/app_manager/templates/app_manager/import_app.html
@@ -1,49 +1,11 @@
 {% extends "app_manager/apps_base.html" %}
 {% load url from future %}
 {% load i18n %}
+{% load hq_shared_tags %}
 
 {% block js-inline %}{{ block.super }}
+    <script src="{% static 'hqwebapp/js/lib/compression.js' %}"></script>
     <script>
-        // http://rosettacode.org/wiki/LZW_compression#JavaScript
-        var LZW = {
-            compress: function (uncompressed) {
-                "use strict";
-                // Build the dictionary.
-                var i,
-                    dictionary = {},
-                    c,
-                    wc,
-                    w = "",
-                    result = [],
-                    dictSize = 256;
-                for (i = 0; i < 256; i += 1) {
-                    dictionary[String.fromCharCode(i)] = i;
-                }
-
-                for (i = 0; i < uncompressed.length; i += 1) {
-                    c = uncompressed.charAt(i);
-                    wc = w + c;
-                    //Do not use dictionary[wc] because javascript arrays
-                    //will return values for array['pop'], array['push'] etc
-                   // if (dictionary[wc]) {
-                    if (dictionary.hasOwnProperty(wc)) {
-                        w = wc;
-                    } else {
-                        result.push(dictionary[w]);
-                        // Add wc to the dictionary.
-                        dictionary[wc] = dictSize++;
-                        w = String(c);
-                    }
-                }
-
-                // Output the code for w.
-                if (w !== "") {
-                    result.push(dictionary[w]);
-                }
-                return result;
-            }
-        };
-
         $(function () {
             $(".historyBack").click(function () {
                 history.back();
