+++ corehq/apps/app_manager/views.py
@@ -35,6 +35,7 @@ from corehq.apps.app_manager.util import save_xform, get_settings_values
 from corehq.apps.domain.models import Domain
 from corehq.apps.domain.views import DomainViewMixin
 from corehq.apps.translations import system_text as st_trans
+from corehq.util.compression import decompress
 from couchexport.export import FormattedRow, export_raw
 from couchexport.models import Format
 from couchexport.shortcuts import export_response
@@ -223,33 +224,6 @@ def form_casexml(req, domain, form_unique_id):
 def app_source(req, domain, app_id):
     app = get_app(domain, app_id)
     return HttpResponse(app.export_json())
-
-
-# http://rosettacode.org/wiki/LZW_compression#Python
-def decompress(compressed):
-    """Decompress a list of output ks to a string."""
-
-    # Build the dictionary.
-    dict_size = 256
-    dictionary = dict((chr(i), chr(i)) for i in xrange(dict_size))
-    # in Python 3: dictionary = {chr(i): chr(i) for i in range(dict_size)}
-
-    w = result = compressed.pop(0)
-    for k in compressed:
-        if k in dictionary:
-            entry = dictionary[k]
-        elif k == dict_size:
-            entry = w + w[0]
-        else:
-            raise ValueError('Bad compressed k: %s' % k)
-        result += entry
-
-        # Add w+entry[0] to the dictionary.
-        dictionary[dict_size] = w + entry[0]
-        dict_size += 1
-
-        w = entry
-    return result
 
 @login_and_domain_required
 def import_app(req, domain, template="app_manager/import_app.html"):
